Link to shipped zlib

+++ SixTrack/buildLibraries.sh
@@ -56,7 +56,31 @@ cd ..
 ####################################
 ### libArchive #####################
 ####################################
-rm -rf libarchive_build
+
+#Support library zlib
+cd zlib
+ZLIB_BASE=$(pwd)
+if [[ -d build ]]; then
+   rm -rf build
+fi
+mkdir build
+if [[ -d install ]]; then
+   rm -rf install
+fi
+mkdir install
+
+cd build
+cmake .. -DCMAKE_INSTALL_PREFIX=$ZLIB_BASE/install -DCMAKE_C_FLAGS=-fPIC
+make
+make install
+cd ..
+
+cd ..
+
+#Then: LibArchive itself
+if [[ -d libarchive_build ]]; then
+    rm -rf libarchive_build
+fi
 mkdir libarchive_build
 cd libarchive_build
 
@@ -78,7 +102,7 @@ if [[ $(uname) == MINGW* ]]; then
     
 else
     #Normal (i.e. UNIX) build.
-    cmake -DCMAKE_BUILD_TYPE=Release -DENABLE_BZip2=OFF -DENABLE_ZLIB=ON -DENABLE_CAT=OFF -DENABLE_CPIO=OFF -DENABLE_EXPAT=OFF -DENABLE_INSTALL=OFF -DENABLE_LIBXML2=OFF -DENABLE_LZMA=OFF -DENABLE_NETTLE=OFF -DENABLE_OPENSSL=OFF -DENABLE_TAR=OFF -DENABLE_CNG=OFF -DENABLE_ICONV=OFF -DENABLE_TEST=OFF -G "Unix Makefiles" ../libarchive -LH
+    cmake -DCMAKE_BUILD_TYPE=Release -DENABLE_BZip2=OFF -DENABLE_ZLIB=ON -DENABLE_CAT=OFF -DENABLE_CPIO=OFF -DENABLE_EXPAT=OFF -DENABLE_INSTALL=OFF -DENABLE_LIBXML2=OFF -DENABLE_LZMA=OFF -DENABLE_NETTLE=OFF -DENABLE_OPENSSL=OFF -DENABLE_TAR=OFF -DENABLE_CNG=OFF -DENABLE_ICONV=OFF -DENABLE_TEST=OFF -DZLIB_LIBRARY=$ZLIB_BASE/install/lib/libz.a -G "Unix Makefiles" ../libarchive -LH
 fi
 
 if [[ $(pwd) == /afs/* ]]; then
